+++ packages/fiori-freestyle-writer/src/index.ts
@@ -1,6 +1,7 @@
 import { join } from 'path';
 import type { Editor } from 'mem-fs-editor';
 import { render } from 'ejs';
+import { ServiceType } from '@sap-ux/odata-service-writer';
 import type { Package } from '@sap-ux/ui5-application-writer';
 import { generate as generateUi5Project } from '@sap-ux/ui5-application-writer';
 import { generate as addOdataService } from '@sap-ux/odata-service-writer';
@@ -87,7 +88,8 @@ async function generate<T>(basePath: string, data: FreestyleApp<T>, fs?: Editor)
             flpAppId: ffApp.app.flpAppId,
             startFile: data?.app?.startFile,
             localStartFile: data?.app?.localStartFile,
-            generateIndex: ffApp.appOptions?.generateIndex
+            generateIndex: ffApp.appOptions?.generateIndex,
+            isCapApplication: ffApp.service?.type === ServiceType.CDS
         })
     };
 
@@ -99,16 +101,20 @@ async function generate<T>(basePath: string, data: FreestyleApp<T>, fs?: Editor)
     } else {
         // Add placeholder middleware so allow adding service later
         const ui5LocalConfigPath = join(basePath, 'ui5-local.yaml');
-        const ui5LocalConfig = await UI5Config.newInstance(fs.read(ui5LocalConfigPath));
-        ui5LocalConfig.addFioriToolsProxydMiddleware({});
-        fs.write(ui5LocalConfigPath, ui5LocalConfig.toString());
+        if(fs.exists(ui5LocalConfigPath)) {
+            const ui5LocalConfig = await UI5Config.newInstance(fs.read(ui5LocalConfigPath));
+            ui5LocalConfig.addFioriToolsProxydMiddleware({});
+            fs.write(ui5LocalConfigPath, ui5LocalConfig.toString());
+        }
     }
 
     // Extend ui5-local.yaml with additional UI5 lib
     const ui5LocalConfigPath = join(basePath, 'ui5-local.yaml');
-    const ui5LocalConfig = await UI5Config.newInstance(fs.read(ui5LocalConfigPath));
-    ui5LocalConfig.addUI5Libs([ushellLib]);
-    fs.write(ui5LocalConfigPath, ui5LocalConfig.toString());
+    if(fs.exists(ui5LocalConfigPath)) {
+        const ui5LocalConfig = await UI5Config.newInstance(fs.read(ui5LocalConfigPath));
+        ui5LocalConfig.addUI5Libs([ushellLib]);
+        fs.write(ui5LocalConfigPath, ui5LocalConfig.toString());
+    }
 
     return fs;
 }
